rename component<->container
