+++ src/client/components/profile/ProfilePage.container.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import ProfileTopbar from './ProfileTopbar.container';
-import ProfileBelt from './ProfileBelt.component';
+import ProfileTopbar from './ProfileTopbar.component';
+import ProfileBelt from './ProfileBelt.container';
 import ProfileCreateTaste from './ProfileCreateTaste.component';
 import ArchetypeTooltip from './ArchetypeTooltip.component';
 import AuthorTooltip from './AuthorTooltip.component';
